Address review nits: remove GET_GROUPS/SYCL_NUM_THREADS, use ceil_div, fix types and formatting
- Remove unused GET_GROUPS function and SYCL_NUM_THREADS constant
- Replace manual ceil-div with at::ceil_div in xpuKernelLoopGroupRange
- Change global_range from uint32_t to int64_t in AveragePool2dKernels.cpp
- Replace SYCL_NUM_THREADS with literal 1024 in LossNLLKernel.cpp
- Fix clang-format issues in DepthwiseConv3dKernels.cpp and MaxUnpoolingKernels.cpp

+++ src/ATen/native/xpu/sycl/AveragePool2dKernels.cpp
@@ -257,7 +257,7 @@ void launch_avg_pool2d_channels_last_kernel(
 
   auto& queue = at::xpu::getCurrentSYCLQueue();
   const uint32_t group_size = static_cast<int>(syclMaxWorkItemsPerSubSlice());
-  const uint32_t global_range =
+  const int64_t global_range =
       xpuKernelLoopGroupRange(total_elements, group_size) * group_size;
 
   auto kfn = AvgPool2dChannelsLastKernelFunctor<scalar_t, accscalar_t, index_t>(
@@ -305,7 +305,7 @@ void launch_avg_pool2d_kernel(
 
   auto& queue = at::xpu::getCurrentSYCLQueue();
   const uint32_t group_size = static_cast<int>(syclMaxWorkItemsPerSubSlice());
-  const uint32_t global_range =
+  const int64_t global_range =
       xpuKernelLoopGroupRange(total_elements, group_size) * group_size;
 
   auto kfn = AvgPool2dKernelFunctor<scalar_t, accscalar_t, index_t>(
@@ -564,7 +564,7 @@ void launch_avg_pool2d_backward_channels_last_kernel(
 
   auto& queue = at::xpu::getCurrentSYCLQueue();
   const uint32_t group_size = static_cast<int>(syclMaxWorkItemsPerSubSlice());
-  const uint32_t global_range =
+  const int64_t global_range =
       xpuKernelLoopGroupRange(total_elements, group_size) * group_size;
 
   auto kfn = AvgPool2dChannelsLastBackwardKernelFunctor<
@@ -615,7 +615,7 @@ void launch_avg_pool2d_backward_kernel(
 
   auto& queue = at::xpu::getCurrentSYCLQueue();
   const uint32_t group_size = static_cast<int>(syclMaxWorkItemsPerSubSlice());
-  const uint32_t global_range =
+  const int64_t global_range =
       xpuKernelLoopGroupRange(total_elements, group_size) * group_size;
 
   auto kfn = AvgPool2dBackwarKernelFunctor<scalar_t, accscalar_t, index_t>(
